+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Upcoming
 
 * Adds a checkbox to the confirm your bid screen - yuki24
+* Adds the ability to style the text in the inverted button with a spinner - yuki24
 
 ### 1.5.1